Organize post sources in same dirs as desired URL layout
This is the simplest way to preserve the original site's
YYYY/MM/title.html pattern. Clean sheet, I'd do flat
YYYY-MM-title.html. But one the one hand I don't want to deal with a
bunch of redirects, and on the other hand the Makefile is saner if the
source markdown, cache, and html files share the same dir structure.